stod partway

+++ CIS22C_FinalProject/CIS22C_FinalProject/Refresh.cpp
@@ -130,19 +130,19 @@ void Refresh::findfunctions()
 				count++;
 			}
 			if (answer[0] == "add") {
-				sheet->setCellData(i, p, to_string(add(stoi(answer[1]), stoi(answer[2]), stoi(answer[3]), stoi(answer[4]))));
+				sheet->setCellData(i, p, to_string(add(stod(answer[1]), stod(answer[2]), stod(answer[3]), stod(answer[4]))));
 			}
 			if (answer[0] == "multiply") {
-				sheet->setCellData(i, p, to_string(multiply(stoi(answer[1]), stoi(answer[2]), stoi(answer[3]), stoi(answer[4]))));
+				sheet->setCellData(i, p, to_string(multiply(stod(answer[1]), stod(answer[2]), stod(answer[3]), stod(answer[4]))));
 			}
 			if (answer[0] == "average") {
-				sheet->setCellData(i, p, to_string(average(stoi(answer[1]), stoi(answer[2]), stoi(answer[3]), stoi(answer[4]))));
+				sheet->setCellData(i, p, to_string(average(stod(answer[1]), stod(answer[2]), stod(answer[3]), stod(answer[4]))));
 			}
 			if (answer[0] == "subtract") {
-				sheet->setCellData(i, p, to_string(subtract(stoi(answer[1]), stoi(answer[2]), stoi(answer[3]), stoi(answer[4]))));
+				sheet->setCellData(i, p, to_string(subtract(stod(answer[1]), stod(answer[2]), stod(answer[3]), stod(answer[4]))));
 			}
 			if (answer[0] == "divide") {
-				sheet->setCellData(i, p, to_string(divide(stoi(answer[1]), stoi(answer[2]), stoi(answer[3]), stoi(answer[4]))));
+				sheet->setCellData(i, p, to_string(divide(stod(answer[1]), stod(answer[2]), stod(answer[3]), stod(answer[4]))));
 			}
 		}
 	}
